s3-trigger/requirements.txt
-e .
aws-cdk.core==1.103.0
aws-cdk.aws-dynamodb==1.103.0
aws_cdk.aws-iam==1.103.0
aws-cdk.aws-lambda==1.103.0
aws-cdk.aws-s3==1.103.0
aws-cdk.aws_s3_notifications==1.103.0
